added basic tasks for cloudControl worker
* added CloudControlBaseTask
* added InterruptableTask
* added LoopTask

The InterruptableTask is an abstract class which provides clean support for a defined interrupt.
The CloudControlBaseTask adds support for the return codes specific to the worker API of cloudControl.
The LoopTask implements an infinite loop over the execution of a task.

The UselessLoopTask is an example of a simple looping task.