added basic cron support
* re-wrote install file
* attached CronParser to Crontab
* added sfCloudControl, an API wrapper for phpcclib
* added sfCloudControlConfiguration, a ValueObject holding configuration for phpcclib
* added CronTask and ReloadCronTask
* secured sfCCCrontab

The CronParser will be attached and re-attached whenever a Crontab instance is saved with changes to the schedule or the instance is being hydrated.

The CronTask checks each Crontab whether it is due to execution and adds a worker if required.
The ReloadCronTask is capable of forcing the CronTask to reload its schedule information while it is running.