adapted CronTask and ReloadCronTask without PID
* added some worker related wrapper methods to sfCloudControl
** addWorker
** removeWorker
** getWorkerList
** getWorkerDetails
** getRunningWorkers
* moved cloudControl wrapper into CloudControlBaseTask

There is no writable filesystem on cloudControl anymore. The PID is not stored anymore.
This also eliminates the issue, that the PID might actually not exist on the current node.

The ReloadCronTask is now a LoopTask.
It does nothing and will be removed by the CronTask after reloading the schedule.